+++ apps/router-e2e/__e2e__/server-loader/app/index.tsx
@@ -18,6 +18,7 @@ export default function Index() {
         <SiteLink href="/second">Go to Second</SiteLink>
         <SiteLink href="/env">Go to Env</SiteLink>
         <SiteLink href="/request">Go to Request</SiteLink>
+        <SiteLink href="/response">Go to Response</SiteLink>
         <SiteLink href="/posts/static-post-1">Go to static Post 1</SiteLink>
         <SiteLink href="/posts/static-post-2">Go to static Post 2</SiteLink>
       </SiteLinks>